+++ src/main/java/org/jenkinsci/plugins/updatebot/ImportGithubRepoBuild.java
@@ -17,57 +17,44 @@
 package org.jenkinsci.plugins.updatebot;
 
 import com.cloudbees.hudson.plugins.folder.Folder;
-import hudson.model.Action;
 import hudson.model.Build;
 import hudson.model.BuildListener;
 import hudson.model.Cause;
-import hudson.model.CauseAction;
 import hudson.model.Item;
-import hudson.model.ItemGroup;
 import hudson.model.ParameterValue;
 import hudson.model.ParametersAction;
 import hudson.model.Result;
-import hudson.plugins.git.RevisionParameterAction;
 import hudson.util.PersistedList;
 import io.fabric8.updatebot.Configuration;
+import io.fabric8.updatebot.UpdateBot;
+import io.fabric8.updatebot.commands.CommandSupport;
 import io.fabric8.updatebot.commands.EnableFabric8;
+import io.fabric8.updatebot.commands.StatusInfo;
 import io.fabric8.utils.Strings;
 import jenkins.branch.BranchProperty;
-import jenkins.branch.BranchPropertyStrategy;
 import jenkins.branch.BranchSource;
 import jenkins.branch.DefaultBranchPropertyStrategy;
 import jenkins.model.Jenkins;
-import jenkins.model.ParameterizedJobMixIn;
-import jenkins.scm.api.SCMRevision;
-import jenkins.scm.api.mixin.ChangeRequestSCMHead2;
-import jenkins.scm.api.trait.SCMHeadAuthority;
 import jenkins.scm.api.trait.SCMSourceTrait;
 import jenkins.scm.impl.trait.RegexSCMHeadFilterTrait;
-import jenkins.scm.impl.trait.WildcardSCMHeadFilterTrait;
 import org.jenkinsci.plugins.github_branch_source.BranchDiscoveryTrait;
 import org.jenkinsci.plugins.github_branch_source.ForkPullRequestDiscoveryTrait;
 import org.jenkinsci.plugins.github_branch_source.GitHubSCMSource;
-import org.jenkinsci.plugins.github_branch_source.GitHubSCMSourceRequest;
 import org.jenkinsci.plugins.github_branch_source.OriginPullRequestDiscoveryTrait;
 import org.jenkinsci.plugins.updatebot.support.JenkinsHelpers;
 import org.jenkinsci.plugins.workflow.multibranch.WorkflowMultiBranchProject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import jenkins.scm.api.SCMSourceOwner;
-
-import jenkins.plugins.git.AbstractGitSCMSource;
-
 import javax.annotation.Nonnull;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.net.URISyntaxException;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
+import java.util.Map;
 
-import static java.util.logging.Level.SEVERE;
 import static org.jenkinsci.plugins.updatebot.UpdateBotPushStepExecution.configureFromGlobalPluginConfiguration;
 
 /**
@@ -109,7 +96,6 @@ public class ImportGithubRepoBuild extends Build<ImportGithubRepoProject, Import
                 for (ParameterValue parameter : allParameters) {
                     String name = parameter.getName();
                     Object value = parameter.getValue();
-                    logger.println("Found " + name + " = " + value);
 
                     if (value != null) {
                         switch (name) {
@@ -126,11 +112,17 @@ public class ImportGithubRepoBuild extends Build<ImportGithubRepoProject, Import
                 }
             }
         }
+        configuration.logCommand(LOG, "importing github repository " + repository + " with pipeline " + pipeline);
 
         if (Strings.isNullOrBlank(repository)) {
             configuration.error(LOG, "No repository parameter found!");
             return Result.FAILURE;
         }
+
+        // remove whitespace or / prefixes or suffixes in case they are copy/paste issues from URL fragments ;)
+        repository = Strings.stripPrefix(Strings.stripSuffix(repository.trim(), "/"), "/");
+        // lets let folks copy/paste a github repo URL too
+        repository = Strings.stripPrefix(repository, "https://github.com/");
         try {
             command.setOrganisationAndRepository(repository);
         } catch (Exception e) {
@@ -149,6 +141,41 @@ public class ImportGithubRepoBuild extends Build<ImportGithubRepoProject, Import
             return Result.FAILURE;
         }
 
+        createMultiBranchProject(configuration, repository);
+
+        Result answer = waitForPullRequestMerge(listener, configuration, command);
+        if (answer != null) {
+            return answer;
+        }
+
+        return triggerScanBuild(configuration, repository);
+    }
+
+    protected Result waitForPullRequestMerge(BuildListener listener, Configuration configuration, CommandSupport lastCommand) throws IOException {
+        UpdateBot updatebot = new UpdateBot();
+        updatebot.setConfiguration(configuration);
+        updatebot.setCommand(lastCommand);
+
+        while (true) {
+            try {
+                Map<String, StatusInfo> status = updatebot.poll();
+                if (!StatusInfo.isPending(status)) {
+                    return null;
+                }
+            } catch (IOException e) {
+                configuration.warn(LOG, "Failed to poll PullRequests " + e, e);
+            }
+            // TODO how to detect we should terminate the build??
+            try {
+                Thread.sleep(10000L);
+            } catch (InterruptedException e) {
+                configuration.warn(LOG, "Build terminated: " + e, e);
+                return Result.ABORTED;
+            }
+        }
+    }
+
+    protected void createMultiBranchProject(Configuration configuration, String repository) {
         // lets create a new Multi-Branch build!
         String[] paths = repository.split("/", 2);
         String organisation = paths[0];
@@ -172,8 +199,10 @@ public class ImportGithubRepoBuild extends Build<ImportGithubRepoProject, Import
         branchSource.setStrategy(strategy);
         sourcesList.add(branchSource);
         JenkinsHelpers.createItem(configuration, parent, repository, project, "WorkflowMultiBranchProject for " + repository);
+    }
 
-        // now lets trigger a scan
+    protected Result triggerScanBuild(Configuration configuration, String repository) {
+        Jenkins jenkins = Jenkins.getInstance();
         Item item = jenkins.getItemByFullName(repository);
         if (item instanceof WorkflowMultiBranchProject) {
             WorkflowMultiBranchProject job = (WorkflowMultiBranchProject) item;