+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.0.6
 
+* import git repository waits for any pending PR to be merged
+* import git repository accepts full https://github.com/org/repo URLs
 * allow the Jenkinsfile library git repository to be configurable
 
 ## 1.0.3